Bootstraped. the C header provided is outdated, ABIs are messed Build command: dub build --compiler=ldc gcc -g bootstrap.c $(pkg-config --static --libs glfw3 gl) -Lout/ -Wl,-rpath=out -linochi2d-c -o bootstrap note that inochi2d being depended on should be our own ES 2.0 port

+++ inochi2d.h
@@ -30,8 +30,8 @@ extern "C" {
     void inCleanup();
     void inUpdate();
     void inBlockProtected(void (*func)());
-    void inViewportSet(float width, float height);
-    void inViewportGet(float* width, float* height);
+    void inViewportSet(int width, int height);
+    void inViewportGet(int* width, int* height);
     #ifdef INOCHI2D_GLYES
         void inSceneBegin();
         void inSceneEnd();
